venxtra-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymongo==4.6.1
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
beanie==1.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1
PyPDF2==3.0.1
tabula-py==2.9.0
camelot-py[cv]==0.11.0
groq==0.11.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
psutil==5.9.6
python-magic==0.4.27